move component_set
