+++ packages/flame/lib/src/components/core/component_set.dart
